backend/requirements.txt
alabaster==0.7.13
asgiref==3.6.0
astor==0.8.1
attrs==22.2.0
Babel==2.11.0
bandit==1.7.4
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
darglint==1.8.1
distlib==0.3.6
Django==4.1.7
# django-admin-honeypot @ git+https://github.com/dmpayton/django-admin-honeypot.git@a840496d183df89965eeb33c9a5dd9ea3919dfff
# django-admin-honeypot-updated-2021==1.2.0
django-appconf==1.0.5
django-axes==5.40.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-crontab==0.7.1
django-csp==3.7
django-filter==22.1
django-http-referrer-policy==1.1.1
django-ipware==4.0.2
django-js-asset==2.0.0
django-storages==1.13.2
django-watermark==0.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
docutils==0.18.1
drf-yasg==1.21.5
eradicate==2.1.0
filelock==3.9.0
filetype==1.2.0
flake8==5.0.4
flake8-bandit==4.1.1
flake8-broken-line==0.6.0
flake8-bugbear==23.2.13
flake8-commas==2.1.0
flake8-comprehensions==3.10.1
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.4.0
flake8-isort==6.0.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
imagesize==1.4.1
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kavenegar==1.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mergedeep==1.3.4
openapi-codec==1.3.2
packaging==23.0
pbr==5.11.1
pep8-naming==0.13.3
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.17
pydocstyle==6.3.0
pyflakes==2.5.0
Pygments==2.14.0
PyJWT==1.7.1
pyproject_api==1.5.0
pyrsistent==0.19.3
pyshorteners==1.0.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.28.2
restructuredtext-lint==1.4.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.3
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
stevedore==5.0.0
swagger-spec-validator==3.0.3
tomli==2.0.1
tox==4.4.6
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.19.0
watchdog==2.2.1
whitenoise==6.3.0
PyMySQL==1.0.3